renamed dgeosettlement.h to dgeosettlement.hpp
