+++ applications/GeoMechanicsApplication/external_bindings/kratos_external_bindings.h
@@ -1,4 +1,4 @@
 #include "custom_workflows/dgeoflow.h"
-#include "custom_workflows/dgeosettlement.h"
+#include "custom_workflows/dgeosettlement.hpp"
 #include <iostream>
 #include <stdlib.h>
