+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_workflows/test_dgeosettlement.cpp
@@ -10,7 +10,7 @@
 //  Main authors:    Richard Faasse
 //
 
-#include "custom_workflows/dgeosettlement.h"
+#include "custom_workflows/dgeosettlement.hpp"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
 #include "tests/cpp_tests/stub_input_utility.h"
 #include "tests/cpp_tests/stub_process_info_parser.h"
